cc2 finished and points and compatibility added to the final answer

+++ Etapa2/structs.hpp
@@ -15,7 +15,7 @@ struct question
 {
     string question;
     vector<pair<string, string>> options;
-    float importance;
+    int importance;
 };
 
 #endif
